+++ java/src/main/kotlin/cc/unitmesh/idea/flow/JvmAutoDevFlow.kt
@@ -39,7 +39,7 @@ class JvmAutoDevFlow : DevFlowProvider() {
         kanban: Kanban?,
         aiRunner: LLMProvider,
         component: ChatCodingPanel,
-        project: Project
+        project: Project,
     ) {
         this.kanban = kanban
         this.connector = aiRunner
@@ -211,7 +211,7 @@ class JvmAutoDevFlow : DevFlowProvider() {
     private fun createCode(
         code: String,
         controllerName: String,
-        isNeedCreateController: Boolean
+        isNeedCreateController: Boolean,
     ) {
         when {
             isNeedCreateController || processor.isController(code) -> {
@@ -257,7 +257,7 @@ class JvmAutoDevFlow : DevFlowProvider() {
         targetEndpoint: String,
         clazz: DtClass,
         storyDetail: String,
-        isNewController: Boolean
+        isNewController: Boolean,
     ): String {
         val allModels = processor.modelList().map { it }
         val relevantName = targetEndpoint.replace("Controller", "")
@@ -287,7 +287,7 @@ class JvmAutoDevFlow : DevFlowProvider() {
     private fun executePrompt(promptText: String): String {
         ui.addMessage(promptText, false, promptText)
         // for answer
-        ui.addMessage(AutoDevBundle.message("devti.loading"))
+        ui.addMessage(AutoDevBundle.message("autodev.loading"))
 
         return runBlocking {
             val prompt = connector.stream(promptText, "", ChatActionType.CHAT)