+++ src/main/kotlin/cc/unitmesh/devti/flow/kanban/impl/GitLabIssue.kt
@@ -9,7 +9,6 @@ import org.gitlab4j.api.models.Issue
 class GitLabIssue(private val apiUrl: String, private val personalAccessToken: String, gitlabUrl: String) : Kanban {
     private lateinit var gitLabApi: GitLabApi
 
-    // Initialize the GitLabApi instance
     init {
         initializeGitLabApi(gitlabUrl, personalAccessToken)
     }
@@ -28,14 +27,14 @@ class GitLabIssue(private val apiUrl: String, private val personalAccessToken: S
     }
 
     override fun getStoryById(storyId: String): SimpleStory {
-        val issue: Issue = gitLabApi.issuesApi.getIssue(apiUrl, storyId.toInt())
+        val issue: Issue = gitLabApi.issuesApi.getIssue(apiUrl, storyId.toLong())
         return SimpleStory(issue.iid.toString(), issue.title, issue.description)
     }
 
 
     override fun updateStoryDetail(simpleStory: SimpleStory) {
         // Create a note as a comment on the issue
-        val issue: Issue = gitLabApi.issuesApi.getIssue(apiUrl, simpleStory.id.toInt())
+        val issue: Issue = gitLabApi.issuesApi.getIssue(apiUrl, simpleStory.id.toLong())
         gitLabApi.notesApi.createIssueNote(apiUrl, issue.iid, simpleStory.description)
     }