+++ src/main/kotlin/cc/unitmesh/devti/gui/chat/ChatCodingService.kt
@@ -1,9 +1,10 @@
 package cc.unitmesh.devti.gui.chat
 
+import cc.unitmesh.cf.core.llms.LlmMsg
 import cc.unitmesh.devti.AutoDevBundle
 import cc.unitmesh.devti.LLMCoroutineScope
 import cc.unitmesh.devti.counit.CoUnitPreProcessor
-import cc.unitmesh.devti.llms.LlmProviderFactory
+import cc.unitmesh.devti.llms.LlmFactory
 import cc.unitmesh.devti.parser.PostCodeProcessor
 import cc.unitmesh.devti.provider.ContextPrompter
 import com.intellij.openapi.application.ApplicationManager
@@ -14,7 +15,7 @@ import kotlinx.coroutines.launch
 import org.jetbrains.eval4j.checkNull
 
 class ChatCodingService(var actionType: ChatActionType, val project: Project) {
-    private val llmProviderFactory = LlmProviderFactory()
+    private val llmFactory = LlmFactory()
     private val counitProcessor = project.service<CoUnitPreProcessor>()
 
     val action = actionType.instruction()
@@ -27,7 +28,7 @@ class ChatCodingService(var actionType: ChatActionType, val project: Project) {
     fun handlePromptAndResponse(
         ui: ChatCodingPanel,
         prompter: ContextPrompter,
-        context: ChatContext? = null
+        context: ChatContext? = null,
     ) {
         val requestPrompt = prompter.requestPrompt()
 
@@ -39,7 +40,7 @@ class ChatCodingService(var actionType: ChatActionType, val project: Project) {
         }
 
         ui.addMessage(requestPrompt, true, prompter.displayPrompt())
-        ui.addMessage(AutoDevBundle.message("autodev.assistant.placeholder"))
+        ui.addMessage(AutoDevBundle.message("autodev.loading"))
 
         ApplicationManager.getApplication().executeOnPooledThread {
             val response = this.makeChatBotRequest(requestPrompt, prompter)
@@ -55,6 +56,26 @@ class ChatCodingService(var actionType: ChatActionType, val project: Project) {
 
                     else -> ui.updateMessage(response)
                 }
+            }
+        }
+    }
+
+    fun handleMsgsAndResponse(
+        ui: ChatCodingPanel,
+        messages: List<LlmMsg.ChatMessage>,
+        currentActionType: ChatActionType
+    ) {
+        val requestPrompt = messages.filter { it.role == LlmMsg.ChatRole.User }.joinToString("\n") { it.content }
+        val systemPrompt = messages.filter { it.role == LlmMsg.ChatRole.System }.joinToString("\n") { it.content }
+
+        ui.addMessage(requestPrompt, true, requestPrompt)
+        ui.addMessage(AutoDevBundle.message("autodev.loading"))
+
+        ApplicationManager.getApplication().executeOnPooledThread {
+            val response = llmFactory.create(project).stream(requestPrompt, systemPrompt, currentActionType)
+
+            LLMCoroutineScope.scope(project).launch {
+                ui.updateMessage(response)
             }
         }
     }
@@ -84,7 +105,7 @@ class ChatCodingService(var actionType: ChatActionType, val project: Project) {
         if (prompter.getActionType() != null) {
             currentActionType = prompter.getActionType()!!
         }
-        return llmProviderFactory.connector(project).stream(requestPrompt, systemPrompt, currentActionType)
+        return llmFactory.create(project).stream(requestPrompt, systemPrompt, currentActionType)
     }
 
     private fun getCodeSection(content: String, prefixText: String, suffixText: String): String {
@@ -97,6 +118,6 @@ class ChatCodingService(var actionType: ChatActionType, val project: Project) {
     }
 
     fun clearSession() {
-        llmProviderFactory.connector(project).clearMessage()
+        llmFactory.create(project).clearMessage()
     }
 }