+++ src/main/kotlin/cc/unitmesh/devti/intentions/action/DefaultDocumentationIntention.kt
@@ -1,7 +1,7 @@
 package cc.unitmesh.devti.intentions.action
 
 import cc.unitmesh.devti.AutoDevBundle
-import cc.unitmesh.devti.custom.CustomDocumentationConfig
+import cc.unitmesh.devti.custom.document.CustomDocumentationConfig
 import cc.unitmesh.devti.intentions.action.base.BasedDocumentationIntention
 import cc.unitmesh.devti.provider.LivingDocumentation
 import com.intellij.openapi.editor.Editor
@@ -9,8 +9,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 
 class DefaultDocumentationIntention: BasedDocumentationIntention() {
-    override val config: CustomDocumentationConfig
-        get() = CustomDocumentationConfig.default()
+    override val config: CustomDocumentationConfig = CustomDocumentationConfig.default()
 
     override fun getText(): String = AutoDevBundle.message("intentions.living.documentation.name")