+++ src/main/kotlin/cc/unitmesh/devti/llms/tokenizer/TokenLength.kt
@@ -1,6 +1,0 @@
-package cc.unitmesh.devti.llms.tokenizer
-
-/**
- * config to [cc.unitmesh.devti.settings.AutoDevSettingsStateNew]
- */
-const val LLM_MAX_TOKEN = 8192