+++ src/main/kotlin/cc/unitmesh/devti/settings/AutoDevSettingsConfigurable.kt
@@ -1,5 +1,6 @@
 package cc.unitmesh.devti.settings
 
+import cc.unitmesh.devti.AutoDevBundle
 import com.intellij.openapi.options.Configurable
 import org.jetbrains.annotations.Nls
 import org.jetbrains.annotations.Nullable
@@ -9,29 +10,18 @@ class AutoDevSettingsConfigurable : Configurable {
     private val component: LLMSettingComponent = LLMSettingComponent(AutoDevSettingsStateNew.getInstance())
 
     @Nls(capitalization = Nls.Capitalization.Title)
-    override fun getDisplayName(): String {
-        return "CodeAssistant"
-    }
+    override fun getDisplayName(): String = AutoDevBundle.message("name")
 
-    override fun getPreferredFocusedComponent(): JComponent? {
-        return null
-    }
+    override fun apply() = component.exportSettings(AutoDevSettingsStateNew.getInstance())
+
+    override fun reset() = component.applySettings(AutoDevSettingsStateNew.getInstance())
+    override fun getPreferredFocusedComponent(): JComponent? = null
 
     @Nullable
-    override fun createComponent(): JComponent {
-        return component.panel
-    }
+    override fun createComponent(): JComponent = component.panel
 
     override fun isModified(): Boolean {
         val settings: AutoDevSettingsStateNew = AutoDevSettingsStateNew.getInstance()
         return component.isModified(settings)
-    }
-
-    override fun apply() {
-        component.exportSettings(AutoDevSettingsStateNew.getInstance())
-    }
-
-    override fun reset() {
-        component.applySettings(AutoDevSettingsStateNew.getInstance())
     }
 }